document-mcp/requirements.txt
fastmcp>=2.13.2
openpyxl>=3.1.2
python-docx>=1.1.0
weasyprint>=60.0
